requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7
pika==1.3.2
opencv-python==4.9.0.80
numpy==1.26.3
pillow==10.2.0
ultralytics==8.1.0
paddlepaddle==2.6.0
paddleocr==2.7.3
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
loguru==0.7.2
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
